src/main/java/com/app/cms/dto/converter/UserConverter.java
package com.app.cms.dto.converter;

import com.app.cms.dto.UserDto;
import com.app.cms.entity.User;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Component;

@Component
public class UserConverter implements ObjectConverter<User, UserDto> {

    private final ModelMapper modelMapper;

    public UserConverter(ModelMapper modelMapper) {
        this.modelMapper = modelMapper;
    }

    @Override
    public UserDto toDto(User user) {
        user.setPassword(null);
        user.setPasswordConfirm(null);

        return modelMapper.map(user, UserDto.class);
    }

    @Override
    public User toEntity(UserDto userDto) {
        return modelMapper.map(userDto, User.class);
    }
}
